johnson funcional

+++ src/graph.py
@@ -34,7 +34,7 @@ def readGraph ():
     # Lee aristas
     for i in range(e):
         a,b,w = raw_input().strip().split()
-        a,b,w = float(a),float(b),float(w)
+        a,b,w = int(a),int(b),float(w)
         graph[(a,b)] = w
         
     return graph
